fix(service-effects): close terminal settlement parity gaps
